libc: the MINIX side of aarch64
Stage 1. minix/lib/libc/arch/aarch64 after the pattern of arch/arm: the
system call and IPC stubs, get_bp, read_tsc and the ucontext routines,
plus the generic files that had an #if for every architecture but this
one.

The stubs implement the ABI in <machine/ipcconst.h>: call number in x0,
endpoint in x1, message in x2, trap type as the svc immediate. The kernel
answers in x0 and x1 and restores the rest from the frame, so the stubs
save nothing except where they need a register for a return value; the
stack pointer stays 16-byte aligned throughout because AArch64 faults on a
misaligned one.

read_tsc reads CNTVCT_EL0, the system counter, not the PMU cycle counter
that earm reads. The cycle counter runs at the core clock, stops in sleep
and is not synchronised between cores; the system counter has one
frequency for the whole machine and never stops. That is what "invariant
TSC" means on x86 and it is the only one of the two that stays true under
SMP, which is the point of this port. The kernel has to set
CNTKCTL_EL1.EL0VCTEN for it to be readable from user mode.

ucontext.S saves what the procedure call standard makes the callee's
business - x19-x28, fp, sp, lr - and reaches errno through __errno() so
that the same code serves the static and the PIC library. makecontext in
_ucontext.c gets an __aarch64__ branch: eight argument registers, the
rest on the stack rounded to keep SP aligned, the ucp parked in x19 for
ctx_start. Its stack_top pointer becomes uintptr_t * - on the two 32-bit
architectures that is the same type it was, and on LP64 an unsigned int
slot would have held half an argument.

__sigreturn picks sf_scp from the top of the signal frame the kernel
built; brksize is a .quad.

lib/libc/arch/aarch64/Makefile.inc and gen/Makefile.inc get the same
!defined(__MINIX) exclusions arm has: NetBSD's signal trampolines and
ucontext files, which MINIX replaces with its own.

minix/include/minix/const.h and libmthread's allocate.c: the arch
conditionals that named i386 and arm now name aarch64 as well; the code
under them is arch-neutral.

Classification: [порт].

+++ minix/lib/libc/sys/getdents.c
@@ -5,7 +5,11 @@
 #include <string.h>
 #include <dirent.h>
 
-ssize_t getdents(int fd, char *buffer, size_t nbytes)
+/* <dirent.h> declares getdents(2) as returning int, and that is what
+ * _syscall() hands back. It used to say ssize_t here, which passed unnoticed
+ * as long as every MINIX architecture was ILP32 and the two types were the
+ * same; on LP64 the compiler rejects the mismatch. */
+int getdents(int fd, char *buffer, size_t nbytes)
 {
   message m;
